add default modules

+++ src/sdk/core/connection.cc
@@ -59,7 +59,7 @@ bool Connection::valid_handshake_header(StaticHeaders &header, size_t expected_s
   return (
       header.message_size == expected_size &&
           header.header_size == StaticHeaders::TotalSize &&
-          header.type == expected_type &&
+          static_cast<uint8_t>(header.type) == expected_type &&
           header.request_id == 0 &&
           header.ack_id == 0
   );
